Use WP-CLI to generate the website
It's like Phake, but better.
